+++ esp-idf-svc/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Breaking
+- BT: `BtDriver::set_device_name` deprecated and not available with ESP-IDF 6.0+. Use the new `EspGap::set_device_name` instead
+  or the existing `EspBleGap::set_device_name`
 - Implement MQTT outbox limit and get_outbox_size()
 - Added argument `subprotocol_list` to `ws_handler` to allow subprotocols to be supported by WebSockets
 - Thread enhancements (#592). Specifically: